add test coverage for auth config

+++ db/migrations/testing/20190705143327_add_salesforce_id_rename_primary.js
@@ -2,11 +2,17 @@ exports.up = function(knex) {
   return Promise.all([
     knex.schema.hasTable("submissions").then(function(exists) {
       if (exists) {
-        return knex.schema.table("submissions", function(table) {
-          table.uuid("salesforce_id").notNullable().unique;
-          table.renameColumn("submission_id", "id");
-          table.dropColumn("contact_id");
-        });
+        return knex.schema
+          .hasColumn("submissions", "salesforce_id")
+          .then(function(exists) {
+            if (!exists) {
+              return knex.schema.table("submissions", function(table) {
+                table.uuid("salesforce_id").notNullable().unique;
+                table.renameColumn("submission_id", "id");
+                table.dropColumn("contact_id");
+              });
+            }
+          });
       }
     })
   ]);
